Fix out of bound panic when Root is the empty string.
This error occured when the target is specified on item level:
The global Target was set to the server string with an empty root,
because the path itself is treated as being local to the current item.

Rationale behind this: We do not want to make assumption about the
global root, but need to ensure that all items connect to the same
server.

Fixes #91.